src/story-generate-api/requirements.txt
fastapi==0.128.0
openai==2.14.0
pydantic==2.12.5
python-dotenv==1.2.1
uvicorn==0.40.0
